+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 
 - Improve Python typing throughout
+- Fix height calculation in FPCaptionedView
+- In figpack_slides, support slide editing of titles
 
 ## [0.2.35] - 2025-10-22